Fix init speed
During init, set speed to full speed and programm it to stop if no temperature limits
triggers another speed

Problem reached was, a temperature between stop and low. With the hysteris
no speed is set (low is set is a temp is greater than the low limit, stop is
set if both temps are below than the stop limits), the fan speed will
alway stay set to high.